refactor: NSAttributeProtocol 생성 및 적용
